Fix the save/remove button for a given user
This required the introduction of a dictionary with recipe's ids as keys
and a boolean as values to tell whether to show the given recipe button
as remove or as save depending on whether it was True or False i.e:
whether the user had already saved the recipe or not.

In order to access a dictionary using a given key, a custom template tag
was added.

Signed-off-by: Carolina Romero Carrasco <mcarolinarc28@gmail.com>